+++ handlers/salesforce_object_retrieve_v1/handler/init.rb
@@ -1,55 +1,55 @@
 # Require the dependencies file to load the vendor libraries
-require File.expand_path(File.join(File.dirname(__FILE__), 'dependencies'))
+require File.expand_path(File.join(File.dirname(__FILE__), "dependencies"))
 
 class SalesforceObjectRetrieveV1
-
   def initialize(input)
     # Set the input document attribute
     @input_document = REXML::Document.new(input)
 
     # Store the info values in a Hash of info names to values.
     @info_values = {}
-    REXML::XPath.each(@input_document,"/handler/infos/info") { |item|
-      @info_values[item.attributes['name']] = item.text
+    REXML::XPath.each(@input_document, "/handler/infos/info") { |item|
+      @info_values[item.attributes["name"]] = item.text
     }
 
     # Retrieve all of the handler parameters and store them in a hash attribute
     # named @parameters.
     @parameters = {}
-    REXML::XPath.match(@input_document, 'handler/parameters/parameter').each do |node|
-      @parameters[node.attribute('name').value] = node.text.to_s
+    REXML::XPath.match(@input_document, "handler/parameters/parameter").each do |node|
+      @parameters[node.attribute("name").value] = node.text.to_s
     end
 
-    @enable_debug_logging = @info_values['enable_debug_logging'].downcase == 'yes' ||
-                            @info_values['enable_debug_logging'].downcase == 'true'
+    @enable_debug_logging = @info_values["enable_debug_logging"].downcase == "yes" ||
+                            @info_values["enable_debug_logging"].downcase == "true"
     puts "Parameters: #{@parameters.inspect}" if @enable_debug_logging
 
     @api_version = "37.0"
   end
 
   def execute()
-    raise "Id cannot be left blank" if @parameters['id'].to_s.empty?
+    raise "Id cannot be left blank" if @parameters["id"].to_s.empty?
 
     auth_info = authorize(
-      @info_values['username'],
-      @info_values['password'],
-      @info_values['security_token'],
-      @info_values['client_id'],
-      @info_values['client_secret']
+      @info_values["username"],
+      @info_values["password"],
+      @info_values["security_token"],
+      @info_values["client_id"],
+      @info_values["client_secret"],
+      false
     )
-    access_token = auth_info['access_token']
-    salesforce_instance = auth_info['instance_url']
+    access_token = auth_info["access_token"]
+    salesforce_instance = auth_info["instance_url"]
 
     ## Get the salesforce contact information associated with the contact id
-    puts "Attempting to retrieve the #{@parameters['type']} information for '#{@parameters['id']}'" if @enable_debug_logging
+    puts "Attempting to retrieve the #{@parameters["type"]} information for '#{@parameters["id"]}'" if @enable_debug_logging
     begin
-      resp = RestClient.get("#{salesforce_instance}/services/data/v#{@api_version}/sobjects/#{@parameters['type']}/#{@parameters['id']}",
-        :content_type => "application/x-www-form-urlencoded", :accept => :json, :authorization => "OAuth #{access_token}")
+      resp = RestClient.get("#{salesforce_instance}/services/data/v#{@api_version}/sobjects/#{@parameters["type"]}/#{@parameters["id"]}",
+                            :content_type => "application/x-www-form-urlencoded", :accept => :json, :authorization => "OAuth #{access_token}")
     rescue RestClient::BadRequest => error
       puts "Error encountered while attempting to retrieve the id:\n#{error.http_body}"
       raise error.http_body
     rescue RestClient::ResourceNotFound => error
-      puts "Not found exception encountered while attempting to retrieve id. Most likely (but not only) reason for this to happen is the #{@parameters['type']} '#{@parameters['id']}' doesn't exist\n#{error.http_body}"
+      puts "Not found exception encountered while attempting to retrieve id. Most likely (but not only) reason for this to happen is the #{@parameters["type"]} '#{@parameters["id"]}' doesn't exist\n#{error.http_body}"
       raise error.http_body
     rescue RestClient::Exception => error
       raise error.http_body
@@ -57,7 +57,7 @@ class SalesforceObjectRetrieveV1
 
     json = JSON.parse(resp)
     raise json["errors"].inspect if json["success"] == false
-    puts "The #{@parameters['type']} '#{@parameters['id']}' was found" if @enable_debug_logging
+    puts "The #{@parameters["type"]} '#{@parameters["id"]}' was found" if @enable_debug_logging
 
     # Return results which is an XML String
     <<-RESULTS
@@ -67,27 +67,33 @@ class SalesforceObjectRetrieveV1
     RESULTS
   end
 
-  def authorize(username,password,security_token,client_id,client_secret)
+  def authorize(username, password, security_token, client_id, client_secret, sandbox)
     # Setting up the hash map that will be turned into json and passed to
     # Salesforce to authenticate
     params = {
       :username => username,
-      :password => password+security_token,
+      :password => password + security_token,
       :client_id => client_id,
       :client_secret => client_secret,
-      :grant_type => "password"
+      :grant_type => "password",
     }
     begin
       resp = RestClient.post(
-        "https://login.salesforce.com/services/oauth2/token",
+        "https://#{sandbox ? "test" : "login"}.salesforce.com/services/oauth2/token",
         params,
         :content_type => "application/x-www-form-urlencoded",
-        :accept => :json
+        :accept => :json,
       )
       auth_info = JSON.parse(resp)
     rescue RestClient::BadRequest => error
       error_json = JSON.parse(error.http_body)
-      raise StandardError, error_json['error_description'].to_s
+      # Try again if received the invalid_grant response using the sandbox URL (test.salesforce.com)
+      if !sandbox && error_json["error"] == "invalid_grant"
+        auth_info = authorize(username, password, security_token, client_id, client_secret, true)
+      else
+        puts error.inspect if @enable_debug_logging
+        raise StandardError, error_json["error_description"].to_s
+      end
     end
 
     return auth_info
@@ -103,6 +109,7 @@ class SalesforceObjectRetrieveV1
     # Globally replace characters based on the ESCAPE_CHARACTERS constant
     string.to_s.gsub(/[&"><]/) { |special| ESCAPE_CHARACTERS[special] }
   end
+
   # This is a ruby constant that is used by the escape method
-  ESCAPE_CHARACTERS = {'&'=>'&amp;', '>'=>'&gt;', '<'=>'&lt;', '"' => '&quot;'}
+  ESCAPE_CHARACTERS = { "&" => "&amp;", ">" => "&gt;", "<" => "&lt;", '"' => "&quot;" }
 end