+++ handlers/salesforce_object_create_v1/changelog.md
@@ -1,2 +1,5 @@
 Salesforce Object Create V1 (2017-09-12)
 * Initial version.  See README for details.
+
+Salesforce Object Update V1 (2020-10-28)
+* Added support for sandbox authentication.